+++ CHANGES.md
@@ -1,4 +1,14 @@
 
+## 6.1.0 (2024-10-17)
+
+* add `_tile_matrices_idx: Dict[str, int]` private attribute to improve `matrices` lookup
+* change `xy_bounds()` and `bounds()` methods to avoid calculation duplication
+
+## 6.0.0 (2024-10-16)
+
+* remove `_geographic_crs` private attribute in `TileMatrixSet` model **breaking change**
+* use `crs.geodetic_crs` property as `geographic_crs` **breaking change**
+
 ## 5.4.2 (2024-08-29)
 
 * better handle anti-meridian crossing bbox in `tms.tiles()` (author @ljstrnadiii, https://github.com/developmentseed/morecantile/pull/154)